Move the spa flush to a workflow
